Refactor backend routes and mock API
