+++ backend/src/data/products.ts
@@ -1,0 +1,13 @@
+export const products = [
+  {
+    id: 1,
+    name: "Масляный фильтр ВАЗ 2107",
+    description: "Масляный фильтр для автомобилей ВАЗ 2101–2107.",
+    articleNumber: "OF-2107",
+    price: 750,
+    category: "Фильтры",
+    carModel: "ВАЗ 2107",
+    stockQuantity: 15,
+    imageUrl: "/images/oil-filter.jpg"
+  }
+];
